Rename directory which contains actual package source code from src/ to protobuf_cloud_datastore_translator/ and update affectec code.
