+++ protobuf_cloud_datastore_translator/utils.py
