+++ tox.ini
@@ -23,9 +23,9 @@ deps = -r requirements-test.txt
        mypy
        mypy-protobuf
 commands =
-    flake8 --config ./lint-configs/.flake8 src/ tests/ examples/
-    pylint -E --rcfile=./lint-configs/.pylintrc src/ tests/ examples/
-    mypy --no-incremental --config-file lint-configs/mypy.ini src/ tests/unit/ tests/integration/
+    flake8 --config ./lint-configs/.flake8 protobuf_cloud_datastore_translator/ tests/ examples/
+    pylint -E --rcfile=./lint-configs/.pylintrc protobuf_cloud_datastore_translator/ tests/ examples/
+    mypy --no-incremental --config-file lint-configs/mypy.ini protobuf_cloud_datastore_translator/ tests/unit/ tests/integration/
 
 [testenv:mypy]
 deps = -r requirements-test.txt
@@ -33,7 +33,7 @@ deps = -r requirements-test.txt
        mypy
        mypy-protobuf
 commands =
-    mypy --no-incremental --config-file lint-configs/mypy.ini src/ tests/unit/ tests/integration/ examples/
+    mypy --no-incremental --config-file lint-configs/mypy.ini protobuf_cloud_datastore_translator/ tests/unit/ tests/integration/ examples/
 
 [testenv:py3.7-integration-tests]
 commands =
